backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
openai==1.76.0
python-dotenv==1.1.0
Flask-SocketIO==5.5.1